+++ CHANGELOG.md
@@ -17,6 +17,8 @@ Note that the versions "0.x.0" probably will include breaking changes.
 - Typir-Langium: new API to register validations to the `$type` of the `AstNode` to validate,
   e.g. `addValidationsRulesForAstNodes({ ReturnStatement: <ValidationRule1>, VariableDeclaration: <ValidationRule2>, ... })`, see (L)OX for some examples
 - Associate inference rules with language keys for an improved performance
+- Typir-Langium: new API to register inference rules to the `$type` of the `AstNode` to validate,
+  e.g. `addInferenceRulesForAstNodes({ MemberCall: <InferenceRule1>, VariableDeclaration: <InferenceRule2>, ...})`, see (L)OX for some examples
 - Thanks to the new chaining API for defining types (see corresponding breaking changes below), they can be annotated in TypeScript-type-safe way with multiple inference rules for the same purpose.
 
 ### Breaking changes